refactor: 점주, 손님 테이블 통합으로 로그인 기능 수정 - SignInParam -> SignInDto로 변경 - 인증 과정에서 Merchant와 관련된 루틴 삭제
