+++ src/main/java/fingerorder/webapp/service/UserService.java
@@ -1,7 +1,7 @@
 package fingerorder.webapp.service;
 
 import fingerorder.webapp.dto.UserDto;
-import fingerorder.webapp.parameter.SignInParam;
+import fingerorder.webapp.dto.SignInDto;
 import fingerorder.webapp.dto.SignUpDto;
 import fingerorder.webapp.parameter.UserEditParam;
 import fingerorder.webapp.parameter.UserParam;
@@ -11,8 +11,8 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 
 public interface UserService extends UserDetailsService {
 	UserDto signUp(SignUpDto signUpParam);
-	UserDto authenticate(SignInParam signInParam);
+	UserDto authenticate(SignInDto signInParam);
 	UserDto getUserInfo(UserParam userParam);
-	List<String> getRoles(SignInParam signInParam);
+	List<String> getRoles(SignInDto signInParam);
 	UserDto editUserInfo(UserEditParam userEditParam);
 }
